requirements.txt
fastapi[all]
groq
httpx
openai
python-dotenv